# Optimized math and advanced examples