--- a/src/play_cljc/macros_common.cljs
+++ b/src/play_cljc/macros_common.cljs
@@ -1,0 +1,1 @@
+(ns play-cljc.macros-common)