On my machine the advanced 3d examples (balls-3d for example) run at 2 fps in Firefox. With these optimizations I got it up to 12 fps, which looks a lot better. They also improve the performance of the native version, but I didn't measure the frame rate.

The most important changes come in the first commit, where I made two changes. First, some profiling of the native balls-3d example (with 500 instead of 50 spheres) revealed that `GLFW/glfwGetFramebufferSize` is shockingly expensive to call. So I refactored the example code to call it only once per frame instead of twice per object per frame. Secondly, I converted the indices data in the shape entity to an array so that the `set-buffers` function doesn't repeatedly converted it to an array once per object per frame. That had a significant impact on both native and web performance, bringing the frame rate of balls-3d up to 6 fps in Firefox.

The rest of the changes I made were to `multiply-matrices` and `inverse-matrix`. The idea is that since we're only dealing with 2x2, 3x3, and 4x4 matrices, it's a lot faster to have branchless functions specifically for those sizes instead of having functions that work for arbitrary sizes using a lot of loops and conditionals. But instead of hand writing 6 different repetitive functions, I wrote two macros, one to produce matrix multiplication functions, and one to produce matrix inversion functions. You just tell them the matrix size and they produce function code for that size. That way, all the loops and logic get run at compile time so that at runtime all that needs to be done is some floating point math. Based on benchmarking done on the JVM using Criterium, the new `multiply-matrices` function is more than 10x faster than the old one and the new `inverse-matrix` function is more than 20x faster than the old one. On the web version, these changes brought the balls-3d frame rate up to 12 fps from 6 fps.

Let me know if any adjustments need to be made. I think there's still room for improvement (for comparison, [https://webgl2fundamentals.org/](https://webgl2fundamentals.org/) runs at 20 fps on my machine in Firefox) so I will probably put up a draft PR soon discussing some more ideas.