@oakes 2022-07-01
This is fantastic work. The matrix macro is a great idea. I'm on a work trip right now but i will look over it and merge when i have time.